Done second task in contacts.js file
In the contacts.js file, import the fs (in the version that works with promises - fs/promises) and path modules for working with the file system.
Create a contactsPath variable and write the path to the contacts.json file in it. To compile the path, use the methods of the path module.
Add asynchronous functions to work with the contact collection. In functions, use the fs module and its readFile() and writeFile() methods. The corresponding functions must return the necessary data using the return statement. Output to the console in written functions should not be carried out.
Export the created functions.